+++ examples/basic/pages/index.tsx
@@ -1,11 +1,13 @@
-import JsonViewer from '@textea/json-viewer'
-import type { NextPage } from 'next'
+import { TextField } from '@mui/material'
+import {
+  JsonViewer,
+  JsonViewerOnChange,
+  setPath
+} from '@textea/json-viewer'
 import type React from 'react'
 import { useCallback, useState } from 'react'
 
-import type { InteractionProps } from '../../../src/type'
-
-function aPlusB (a:number, b: number) {
+function aPlusB (a: number, b: number) {
   return a + b
 }
 
@@ -25,14 +27,34 @@ const example = {
   date: new Date('Tue Sep 13 2022 14:07:44 GMT-0500 (Central Daylight Time)')
 }
 
-const IndexPage: NextPage = () => {
+const IndexPage: React.FC = () => {
+  const [indent, setIndent] = useState(2)
   const [src, setSrc] = useState<object>(() => example)
-  const handleEdit = useCallback((update: InteractionProps) => {
-    setSrc(update.updated_src)
-  }, [])
   return (
     <div>
-      <JsonViewer src={src} onEdit={handleEdit}/>
+      <TextField
+        value={indent}
+        type='number'
+        onChange={
+          event => {
+            const indent = parseInt(event.target.value)
+            if (indent > -1 && indent < 10) {
+              setIndent(indent)
+            }
+          }
+        }
+      />
+      <JsonViewer
+        value={src}
+        indentWidth={indent}
+        onChange={
+          useCallback<JsonViewerOnChange>(
+            (path, oldValue, newValue) => {
+              setSrc(src => setPath(src, path, newValue))
+            }, []
+          )
+        }
+      />
     </div>
   )
 }
