--- examples/next/next.config.js
@@ -1,15 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  optimizeFonts: true,
-  experimental: {
-    externalDir: true
-  }
-}
-
-const withTM = require('next-transpile-modules')([], {
-  resolveSymlinks: true,
-  debug: false
-})
-
-module.exports = withTM(nextConfig)
