--- examples/next/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@textea/json-viewer-example-next",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@textea/json-viewer": "workspace:^",
-    "next": "^12.3.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "@textea/dev-kit": "^0.12.16",
-    "@types/node": "^18.7.17",
-    "@types/react": "^18.0.19",
-    "@types/react-dom": "^18.0.6",
-    "next-transpile-modules": "^9.0.0",
-    "typescript": "^4.8.3"
-  }
-}
