--- examples/next/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "extends": "@textea/dev-kit/config/tsconfig.root.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "@textea/json-viewer": ["../../src/index"]
-    },
-    "jsx": "preserve",
-    "jsxImportSource": "react",
-    "noEmit": true
-  },
-  "include": [
-    "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
-}
