--- src/components/DataTypes/Boolean.tsx
@@ -1,15 +1,0 @@
-import type React from 'react'
-
-import type { DataTypeProps } from '../../types/data-type'
-// theme
-import Theme from './../../themes/getStyle'
-import { DataTypeLabel } from './DataTypeLabel'
-
-export const JsonBoolean: React.FC<DataTypeProps<boolean>> = (props) => {
-  return (
-    <div {...Theme(props.theme, 'boolean')}>
-      <DataTypeLabel {...props} type_name='bool'/>
-      {props.value ? 'true' : 'false'}
-    </div>
-  )
-}
