--- src/helpers/parseInput.ts
@@ -1,78 +1,0 @@
-function formatResponse (type: string, value: any) {
-  return {
-    type,
-    value
-  }
-}
-
-export function parseInput (input: string) {
-  // following code is to make the best guess at
-  // the type for a variable being submitted.
-
-  // we are working with a serialized data representation
-  input = input.trim()
-  try {
-    input = JSON.stringify(JSON.parse(input))
-    if (input.startsWith('[') && input.endsWith(']')) {
-      // array
-      return formatResponse('array', JSON.parse(input))
-    } else if (input.startsWith('{') && input.endsWith('}')) {
-      // object
-      return formatResponse('object', JSON.parse(input))
-    } else if (
-      input.match(/-?\d+\.\d+/) &&
-            input.match(/-?\d+\.\d+/)?.[0] === input
-    ) {
-      // float
-      return formatResponse('float', parseFloat(input))
-    } else if (
-      input.match(/-?\d+e-\d+/) &&
-            input.match(/-?\d+e-\d+/)?.[0] === input
-    ) {
-      // scientific float
-      return formatResponse('float', Number(input))
-    } else if (
-      input.match(/-?\d+/) &&
-            input.match(/-?\d+/)?.[0] === input
-    ) {
-      // integer
-      return formatResponse('integer', parseInt(input))
-    } else if (
-      input.match(/-?\d+e\+\d+/) &&
-            input.match(/-?\d+e\+\d+/)?.[0] === input
-    ) {
-      // scientific integer
-      return formatResponse('integer', Number(input))
-    }
-  } catch (e) {
-    // no-op
-  }
-
-  // run in case input was not serializable
-  input = input.toLowerCase()
-  switch (input) {
-    case 'undefined': {
-      return formatResponse('undefined', undefined)
-    }
-    case 'nan': {
-      return formatResponse('nan', NaN)
-    }
-    case 'null': {
-      return formatResponse('null', null)
-    }
-    case 'true': {
-      return formatResponse('boolean', true)
-    }
-    case 'false': {
-      return formatResponse('boolean', false)
-    }
-    default: {
-      // check to see if this is a date
-      if (!Number.isNaN(Date.parse(input))) {
-        return formatResponse('date', new Date(input))
-      }
-    }
-  }
-
-  return formatResponse('string', input)
-}
