+++ vite.config.ts
@@ -1,5 +1,6 @@
 import react from '@vitejs/plugin-react'
 import { resolve } from 'path'
+import dts from 'vite-plugin-dts'
 import { defineConfig } from 'vitest/config'
 
 import { peerDependencies } from './package.json'
@@ -25,11 +26,7 @@ export default defineConfig({
     }
   },
   plugins: [
-    react()
-    // dts({
-    //   outputDir: ['dist'],
-    //   insertTypesEntry: true,
-    //   staticImport: true
-    // })
+    react(),
+    dts()
   ]
 })
